Remove deprecated files and update token display components
- Deleted `dev.degenduel.me.har` and `DEWORKSPACE.md` as they are no longer needed.
- Updated `package.json` and `package-lock.json` to reflect the new version of `@branchmanager69/degenduel-shared`.
- Refactored token display components, replacing `EnhancedTokensPage` with `TokensPage` for improved UI consistency.
- Adjusted various components to ensure compatibility with the new token display structure and removed references to deprecated components.

This commit streamlines the codebase by removing obsolete files and enhancing the token display functionality.
