--- DEWORKSPACE.md
@@ -1,45 +1,0 @@
-# Detransitioning From Workspace
-
-Npm workspaces are powerful, but they definitely introduce complexity, especially around the build process, dependency resolution (as you're seeing with `lucide-react` and `mipd`), and CI/CD setups that aren't designed with a monorepo root in mind. If the *primary* goal was just sharing TypeScript types/interfaces, then yes, workspaces might indeed be overkill.
-
-Let's look at the alternatives suggested by that other person:
-
-1.  **TypeScript Project References:**
-    * **Assessment:** This is often the **best and most idiomatic solution** specifically for sharing TypeScript code (especially just types) between related projects *without* treating the shared code as a published npm package.
-    * **How it works:** You'd typically have your `degenduel-shared` directory with its own `tsconfig.json`. Your `degenduel-fe` and `degenduel-be` projects would then have references added to their `tsconfig.json` files pointing to the shared project.
-    * **Pros:** Designed for this exact scenario. Great IDE support (Go To Definition, Find References work across project boundaries). Avoids npm dependency management headaches *for the shared types*. Can simplify build setups if configured correctly (`tsc --build` can build dependencies). Doesn't require publishing packages or complex workspace linking for type checking.
-    * **Cons:** Requires understanding `tsconfig.json` references setup. Build tools (like Vite or Webpack) might need slight configuration to resolve paths correctly if you're importing compiled JS, though often unnecessary if just importing types.
-    * **Verdict:** Highly recommended if type-sharing is the main goal. It directly addresses the problem with less overhead than workspaces.
-
-2.  **Git Submodules:**
-    * **Assessment:** A viable, but often more cumbersome, Git-level solution.
-    * **How it works:** `degenduel-shared` becomes its own Git repo, included within `degenduel-fe` and `degenduel-be` repos via `git submodule add`.
-    * **Pros:** Explicitly versions the shared code via Git commits. CI checkout processes usually support pulling submodules. Keeps repositories separate.
-    * **Cons:** Git submodules have a reputation for being fiddly and adding friction to the developer workflow (updating, committing, potential for detached HEAD states within the submodule). Doesn't inherently integrate with TS/JS build tooling; you still need to reference files via relative paths. Can be confusing for team members unfamiliar with them.
-    * **Verdict:** Works, but often less smooth for JS/TS development workflows compared to TS Project References or even workspaces. Might be overkill and introduce different kinds of complexity.
-
-3.  **Copy-and-sync approach:**
-    * **Assessment:** The simplest conceptually, but the most brittle and error-prone.
-    * **How it works:** Literally copying type definition files between projects, possibly automated with a script.
-    * **Pros:** Very little setup. Projects remain completely independent.
-    * **Cons:** High risk of divergence (forgetting to copy, editing the copy instead of the source). Violates the DRY (Don't Repeat Yourself) principle. Becomes unmanageable quickly as the number of shared types grows. Requires manual discipline or potentially complex scripting to keep reliable.
-    * **Verdict:** Generally discouraged for anything beyond the most trivial sharing needs due to maintenance headaches and potential for inconsistencies.
-
-**Conclusion & Recommendation:**
-
-The assessment you received is spot on. **If your primary need is just sharing TypeScript types and interfaces**, npm workspaces are likely adding unnecessary complexity, especially given your CI setup challenges.
-
-**TypeScript Project References** seem like the ideal alternative. They provide a structured way to share code between TypeScript projects with direct integration into the TS compiler and IDEs, avoiding the need for the shared code to be treated like a full npm dependency with all the associated resolution and linking issues.
-
-**Moving Forward:**
-
-You have a choice:
-
-1.  **Continue fixing the workspace:** We can proceed with installing `mipd`, fixing the build script, and then tackling the `package-lock.json` generation for CI. This gets your current setup working but leaves the underlying workspace complexity.
-2.  **Explore switching:** We could pause fixing the build and explore restructuring to use TypeScript Project References instead. This involves:
-    * Removing the workspace configuration from the root `package.json`.
-    * Setting up `tsconfig.json` files in `degenduel-fe`, `degenduel-be`, and `degenduel-shared` to reference each other appropriately.
-    * Adjusting import paths if necessary.
-    * This would likely simplify your dependency management *for the shared types* and potentially resolve the CI issue by allowing `degenduel-fe` to function more like a standalone project again (without needing a root workspace lock file).
-
-Given your frustration, exploring **Option 2 (TypeScript Project References)** might be a worthwhile investment to reduce long-term complexity, assuming type-sharing is the main driver. What are your thoughts?
