--- dev.degenduel.me.har
@@ -1,11 +1,0 @@
-{
-  "log": {
-    "version": "1.2",
-    "creator": {
-      "name": "WebInspector",
-      "version": "537.36"
-    },
-    "pages": [],
-    "entries": []
-  }
-}
