+++ src/AUTH_README.md
@@ -105,7 +105,7 @@ We have successfully implemented the core components of the new unified authenti
    - ✅ `AuthService.ts` - Central authentication service
    - ✅ `UnifiedAuthContext.tsx` - Unified authentication provider
    - ✅ `UnifiedWebSocketContext.tsx` - WebSocket provider integrated with auth
-   - ✅ `App.unified.tsx` - Reimplemented App component using the new system
+   - ✅ `App.tsx` - Reimplemented App component using the new system
    - ✅ Unified route guards (`AuthenticatedRoute.unified.tsx`, `AdminRoute.unified.tsx`, `SuperAdminRoute.unified.tsx`)
 
 2. **Tests Created:**
@@ -270,7 +270,7 @@ src/
 │       └── SuperAdminRoute.unified.tsx     // SuperAdmin route
 ├── examples/
 │   └── AuthMigrationExample.tsx  // Example showing migration in practice
-└── App.unified.tsx               // App with unified auth
+└── App.tsx               // App with unified auth
 ```
 
 ### Files to Replace/Update
