+++ src/WEBSOCKET_STANDARDIZATION_IMPLEMENTATION.md
@@ -1,0 +1,71 @@
+# WebSocket Standardization Implementation
+
+## Changes Made
+
+We've successfully standardized the WebSocket action names across the frontend codebase to ensure consistent communication with the backend:
+
+1. **Enhanced DDWebSocketActions Enum**:
+   - Expanded the enum with comprehensive action names for all WebSocket topics
+   - Organized actions by topic with clear section headers
+   - Added detailed documentation on implementation patterns
+   - Fixed duplicate identifier issues (renamed `UPDATE` to `TERMINAL_UPDATE` in relevant context)
+
+2. **Updated Terminal Data Hook**:
+   - Modified `useTerminalData.ts` to use the standardized action constants
+   - Fixed the message interface to use proper types
+   - Replaced string literals with enum constants for type safety
+   - Ensured compatibility with backend expectations
+
+3. **Added Documentation**:
+   - Created `WEBSOCKET_ACTION_NAMES.md` with guidance for developers
+   - Added implementation notes in the enum itself
+   - Provided code examples for before/after migration
+
+## The Problem We Solved
+
+The codebase had inconsistencies in WebSocket action naming that caused communication issues:
+- Frontend was using `'GET_TERMINAL_DATA'` while backend expected `'getData'`
+- Some action names were uppercase with underscores, others were camelCase
+- No centralized source of truth for WebSocket action names
+
+## Benefits of This Implementation
+
+1. **Type Safety**: Using enums instead of string literals prevents typos and ensures only valid actions are used
+2. **Consistency**: All WebSocket action naming now follows the same pattern
+3. **Developer Experience**: Clear documentation and organization makes the system easier to understand
+4. **Frontend-Backend Compatibility**: Action string values now match exactly what the backend expects
+5. **Maintainability**: Central location for all action names makes updates simpler
+
+## Usage Example
+
+Before:
+```typescript
+// Inconsistent string literals led to errors
+ws.request(TopicType.TERMINAL, 'GET_TERMINAL_DATA');  // Didn't match backend's 'getData'
+```
+
+After:
+```typescript
+import { DDWebSocketActions } from '../../../websocket-types-implementation';
+
+// Type-safe constants that match backend expectations
+ws.request(TopicType.TERMINAL, DDWebSocketActions.GET_DATA);  // Matches backend's 'getData'
+```
+
+## Future Improvements
+
+1. **Shared Package Integration**: Consider moving all action constants to the `@branchmanager69/degenduel-shared` package to share directly with the backend
+2. **Topic-Specific Action Interfaces**: Create interfaces to restrict which actions can be used with which topics
+3. **Code Generator**: Create a tool to auto-generate action constants from backend handler documentation
+4. **Testing Utilities**: Add test helpers to validate WebSocket messages against expected formats
+
+## Migration Plan
+
+For existing code, we recommend this incremental approach:
+
+1. Identify WebSocket hooks using string literals (grep for `ws.request` and `ws.command`)
+2. Replace string literals with corresponding DDWebSocketActions constants
+3. Run type checking to ensure compatibility
+4. Test functionality with backend to confirm communication works
+
+Recommended timeline: Migrate 2-3 hooks per day, starting with the most critical ones.
