+++ src/contexts/TwitterAuthContext.tsx
@@ -65,7 +65,7 @@ export const TwitterAuthProvider: React.FC<{ children: ReactNode }> = ({ childre
       "1. Replace <TwitterAuthProvider> with <UnifiedAuthProvider> in your app component\n" +
       "2. Remove nested auth provider hierarchy (AuthProvider, PrivyAuthProvider, TwitterAuthProvider)\n" +
       "3. Update usage of useTwitterAuth() to useAuth() from UnifiedAuthContext\n" +
-      "4. See App.unified.tsx for reference implementation\n" +
+      "4. See App.tsx for reference implementation\n" +
       "5. Twitter functions are now accessible via useAuth().linkTwitter()"
     );
   }, []);
