+++ src/main.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 
-import { App } from "./App.unified";
+import { App } from "./App";
 import { AppErrorBoundary } from "./components/shared/AppErrorBoundary";
 import "./index.css";
 import { initializeClientLogForwarder } from "./utils/clientLogForwarder";
