+++ tsconfig.json
@@ -42,5 +42,6 @@
     }
   },
   "include": ["src"],
+  "exclude": ["src/stories/archive/**/*", "archive/**/*"],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
